# Improve Rank Filter Performance