… Changes.md
… python/GafferImageTest/MedianTest.py
… python/GafferImageTest/images/noisyBlobs.exr
… python/GafferImageTest/images/noisyBlobsMed72x67.exr
… src/GafferImage/ImageSampler.cpp
… src/GafferImage/Merge.cpp
… src/GafferImage/RankFilter.cpp