Sorry this took so long, probably should have addressed all the exceptional floating point value stuff separately - wasn't expecting that to open as many cans of worms as it did, but I think it's in a pretty good place now.

The interesting bit is the Median filter optimization, which I've tried to document pretty thoroughly.

The performance improvements are pretty good - my benchmarks look like:
```
3 Performance Improvements :

- testPerf (GafferImageTest.DilateTest.DilateTest) : was 5.72s now 0.08s (99% reduction)
- testPerf (GafferImageTest.ErodeTest.ErodeTest) : was 5.57s now 0.09s (98% reduction)
- testPerf (GafferImageTest.MedianTest.MedianTest) : was 7.11s now 0.66s (91% reduction)
```