@danieldresser-ie 2023-08-05
OK, I think I have now addressed or replied to all comments.

Fortunately the Windows CI failure had a straightforward explanation: I'm using OIIO's noise functions to prepare test data, and it appears that one of those is yielding different results on Windows. There is one test image for large radius median filters that was too slow to regenerate every time with OIIO, so I stored that, but I was generating the source image dynamically using OIIO - the mismatch was between the noise generated source on Linux and Windows. I've worked around this by storing the source image as well ... bit of a shame when the source file is twice as large as the filtered one ( noise doesn't compress well ), but 257K isn't totally out of line with the sizes of other test files, so this is probably a reasonable work around to the OIIO bug.