@johnhaddon 2023-08-07
> OK, I think I have now addressed or replied to all comments.

Thanks - I've resolved all the discussions except for the uncommented test that is now failing. I think that test perhaps needs rewriting and moving to ImageSamplerTest anyway.

> I've worked around this by storing the source image as well ... bit of a shame when the source file is twice as large as the filtered one ( noise doesn't compress well ), but 257K isn't totally out of line with the sizes of other test files, so this is probably a reasonable work around to the OIIO bug.

I really really want to keep the repo size manageable, and GafferImageTest is currently the biggest problem is this regard. Can we at least use half float? And perhaps we could also reduce the resolution by a half?