@danieldresser-ie 2023-08-08
Fixed test, added matching ImageSampleTest so we're testing the code that actually changed, converted stored test images to half float. Should be good to go